# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Unity游戏开发辅助工具API文档",
        "title": "UnityTool API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/andycai/unitool",
            "email": "your-email@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/andycai/unitool/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/stats": {
            "get": {
                "description": "获取游戏性能统计数据列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "获取统计数据列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/stats.StatsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stats/{id}": {
            "get": {
                "description": "获取单条游戏性能统计数据详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "获取统计数据详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "统计数据ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stats.StatsResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的统计数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "删除统计数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "统计数据ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "stats.StatsResponse": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "graphics_gpu": {
                    "type": "string"
                },
                "graphics_mem": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "login_id": {
                    "type": "string"
                },
                "package_name": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "system_cpu": {
                    "type": "string"
                },
                "system_mem": {
                    "type": "number"
                }
            }
        }
    }
}